// fbf-app/src/app/model/drmodel.ts
export class DRModel {
    Id: number;
    DRNumber: string;
    SupplierCustomerName: string;
    SupplierCustomerId: number;
    ReceiptType: number;
    Date: string;
    Note: string;
    Project: string;
    Address: string;
    DeliveredBy: string;
    VehicleNo: string;
    CreatedBy: string;
}